Refactored: Created a more robust directory structure for hub
