+++ src/FlexBrowserApp.ts
@@ -1,7 +1,7 @@
 import { ViewController } from '@Hi/ViewController';
 import BrowserPreferences from './BrowserPreferences';
 import FlexBrowserWindow from './FlexBrowserWindow';
-import BrowserFrameComposer from './FlexHub/BrowserFrameComposer';
+import BrowserFrameComposer from './FlexHub/BrowserFrameComposer/BrowserFrameComposer';
 import FlexHub from './FlexHub/FlexHub';
 import FlexPreferences from './FlexHub/FlexPreferences';
 import FlexWindowsViewer from './FlexHub/FlexWindowsViewer';
