+++ src/FlexHub/BrowserFrameComposer/BrowserFrameCanvas.ts
