+++ src/FlexHub/BrowserFrameComposer/PartitionComponentOrganizer.ts
