+++ src/FlexHub/BrowserFrameComposer/PartitionEditor.ts
