+++ src/FlexHub/BrowserFrameComposer/PartitionerComponent.ts
